% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{edges_to_igraph}
\alias{edges_to_igraph}
\title{Convert edges to igraph object}
\usage{
edges_to_igraph(edges, group1.id = NULL, group2.id = NULL,
  count.id = NULL, group1.cluster = NULL, group2.cluster = NULL)
}
\arguments{
\item{edges}{Matrix or data.frame or igraph of connection strengths as counts}

\item{group1.id}{A character string indicating the name of group1 identifier
variable in the \code{edges} data.frame. It is required in the case of data.frame.}

\item{group2.id}{A character string indicating the name of group2 identifier
variable in the \code{edges} data.frame. It is required in the case of data.frame.}

\item{count.id}{A character string indicating the name of count identifier
variable in the \code{edges} data.frame. The variable must be numeric.}

\item{vertices}{A data frame with vertex metadata, or NULL.}
}
\value{
igraph object
}
\description{
Convert edges to igraph object
}
