% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmem.R
\name{lmem}
\alias{lmem}
\title{Estimates regression params using a normal model for errors, 
a normal model for Beta, and a Cauchy for sigma}
\usage{
lmem(formula, data, b_loc = 0, b_scale = 1000, sigma_scale = 50, ...)
}
\arguments{
\item{formula}{The formula for the regression. Dependent variable ~ independent variables}

\item{data}{The data used for the regression, data frame}

\item{b_loc}{Prior for Beta mean}

\item{b_scale}{Prior for Beta sd}

\item{sigma_scale}{Prior for sigma sd (mean is 0)}
}
\value{
A trained stanmodel object
}
\description{
Estimates regression params using a normal model for errors, 
a normal model for Beta, and a Cauchy for sigma
}
