% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmem.R
\name{lmem}
\alias{lmem}
\title{Estimates a latent space model based on position data}
\usage{
LSNM(edges, D = 2, ...)
}
\arguments{
\item{edges}{Matrix of connection strengths as counts}

\item{D}{The dimensionality of the latent space, 2 dimensions is recommended}

\value{
A trained stanmodel object
}
\description{
Estimates a latent space model based on position data
}
