% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSM.R
\name{plot.LSM}
\alias{plot.LSM}
\title{Plot the posterior means of a LSM object}
\usage{
plot.LSM(LSM_Object, group1_cluster = NULL, group2_cluster = NULL,
  main = "Estimated LSM Positions", legend = c("Group1", "Group2"),
  legend_position = "topleft", ...)
}
\arguments{
\item{LSM_Object}{A trained object of class LSM}

\item{group1_cluster}{A vector representing the cluster of group1}

\item{group2_cluster}{A vector representing the cluster of group2}
}
\value{
a plot of the posterior means
}
\description{
\code{plot.LSM()} takes an object returned by
\code{LSM}, and returns a plot of the posterior means.
Use standard arguments to the \code{plot} function to modify the plot as needed.
}
