% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSNM.R
\name{plot.LSNM}
\alias{plot.LSNM}
\title{Plot the posterior means of a LSNM object}
\usage{
plot.LSNM(LSNM_Object, group1_cluster = NULL, group2_cluster = NULL,
  main = "Estimated LSNM Positions", legend = c("Group1", "Group2"),
  legend_position = "topleft", ...)
}
\arguments{
\item{LSNM_Object}{A trained object of class LSNM}

\item{group1_cluster}{A vector representing the cluster of group1}

\item{group2_cluster}{A vector representing the cluster of group2}
}
\value{
a plot of the posterior means
}
\description{
\code{plot.LSNM()} takes an object returned by
\code{LSNM}, and returns a plot of the posterior means.
Use standard arguments to the \code{plot} function to modify the plot as needed.
}
