\name{polnet-package}
\docType{package}
\alias{polnet-package}
\alias{polnet}
\title{Analysis of Political Networks}
\description{A collection of functions for the analysis of political networks.}
\details{
The package provides useful functions for the analysis of political networks, broadly defined. To display citation information, execute \code{citation("polnet")}.
}
\seealso{
\code{\link{regindep}}
}
\author{
Philip Leifeld (\url{http://www.philipleifeld.de})
}
