% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSNM.R
\name{summary.LSNM}
\alias{summary.LSNM}
\title{Get summaries of a LSNM object}
\usage{
summary.LSNM(LSNM_Object, low_perc = 0.1, high_perc = 0.9)
}
\arguments{
\item{LSNM_Object}{A trained object of class LSNM}

\item{low_perc}{The bottom range of the desired credible interval, defaults to 0.1}

\item{high_perc}{The top range of the credible interval, defaults to 0.9}
}
\value{
A matrix that includes the mean, standard deviation, and credible interval of the latent space estimated by the LSNM algorithm. The row embeddings are the client latent space positions, while the column embeddings are the legislator latent space positions.
}
\description{
\code{summary.LSNM()} takes an object returned by
\code{LSNM}, and returns a matrix of the mean,
standard deviation, and credible interval of the latent space with all chains being merged
}
