% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSM.R
\name{choose.fix}
\alias{choose.fix}
\title{Choose the actors whose coordinates will be fixed}
\usage{
choose.fix(LSM_Object, n.wild = 1, choose.method = "axis")
}
\arguments{
\item{LSM_Object}{A trained object of class LSM}

\item{n.wild}{The number of actors to be fixed in each octants}

\item{method}{The method to choose the wildest actors}
}
\value{
a list containing the parameters of actors whose coordinates are fixed.
}
\description{
\code{choose.fixed()} takes an object returned by
\code{LSM}, and returns a list containing the parameters
of actors whose coordinates are fixed.
}
