\name{polnet-package}
\alias{polnet-package}
\alias{polnet}
\docType{package}
\title{
\packageTitle{polnet}
}
\description{
\packageDescription{polnet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{polnet}
\packageIndices{polnet}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{polnet}

Maintainer: \packageMaintainer{polnet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
